remove acquisition control
