+++ .github/workflows/static-tests.yml
@@ -21,7 +21,6 @@ jobs:
           "exam-manager",
           "workflow-manager",
           "device-manager",
-          "acquisition-control",
           "mri/sequence-manager",
           "mri/recos/cartesian-reco-service",
         ]
