+++ infrastructure/nginx_config.conf
@@ -25,8 +25,4 @@ server {
   location /api/v1/mri/sequences {
     proxy_pass http://sequence-manager:8000/api/v1/mri/sequences;
   }
-
-  location /api/v1/mri/acquisitioncontrol {
-    proxy_pass http://acquisition-control:8000;
-  }
 }
