--- services/acquisition-control/app/__init__.py
@@ -1,1 +1,0 @@
-"""Init file, that enables sphinx to detect this package."""
