--- services/acquisition-control/pyproject.toml
@@ -1,62 +1,0 @@
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
-
-[tool.poetry]
-name = "acquisition_control"
-version = "0.0.1"
-description = "Acquisition control. Receives control cmd from ui and controls scans on devices."
-authors = [
-    "Joahnnes Behrens <johannes.behrens@brain-link.de>"
-]
-readme = "README.md"
-packages = [{include = "app"}]
-
-[tool.poetry.dependencies]
-python = ">=3.10,<3.11"
-fastapi = "^0.96.0"
-requests = "^2.31.0"
-uvicorn = "^0.22.0"
-scanhub = "^0.0.4.dev3"
-types-requests = "^2.31.0.1"
-httpx = "^0.25.0"
-
-[tool.poetry.group.lint.dependencies]
-ruff = "^0.1.11"
-mypy = "^1.8.0"
-
-[tool.poetry.group.test.dependencies]
-coverage = "*"
-codecov = "*"
-pytest = "*"
-
-[tool.poetry.group.docs.dependencies]
-sphinx = "*"
-sphinx_rtd_theme = "*"
-
-[tool.ruff]
-src = ["app"]
-line-length = 120
-indent-width = 4
-target-version = "py310"
-
-[tool.ruff.lint]
-# Enable linters
-# W = pycodestyle, F = Pyflakes, S = flake8-bandit, I = isort, N = pep8-naming, D = pydocstyle
-select = ["W", "F", "S", "I", "N", "D"]
-
-[tool.ruff.lint.flake8-type-checking]
-strict = true
-
-[tool.ruff.lint.pydocstyle]
-convention = "numpy"
-
-[tool.mypy]
-warn_return_any = false
-check_untyped_defs = true
-warn_no_return = true
-warn_unreachable = true
-python_version = "3.10"
-ignore_missing_imports = true
-mypy_path = "app"
-explicit_package_bases = true   # TODO: Fix imports
