+++ tools/device/koma_example.py
@@ -43,4 +43,6 @@ class Mri(Device):
         return
 
 cur_path = pathlib.Path(__file__).parent.resolve()
+
+## TBD: This shall be replaced with the workflow-manager's API
 mri = Mri("127.0.0.1:8080/api/v1/mri/acquisitioncontrol", f"{cur_path}\\records\\", app)
